electron: always fork except if `--no-cluster`
The generated application code is not forking the backend when working
from sources. Running the backend in the Electron main process was only
meant to help with debugging. This commit makes the generated code fork
the backend unless a `--no-cluster` flag is specified.

If you want the generated code to not fork the backend in a sub-process,
you can either pass a `--no-cluster` flag to your application, or set
`THEIA_ELECTRON_NO_BACKEND_FORK=1` environment variable.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>